requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-use-email-as-username==1.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-excel==2.4.1
drf-renderer-xlsx==1.0.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
openpyxl==3.1.5
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
XlsxWriter==3.2.0
xlwt==1.3.0
